mpi/spawn: split MPIR implementations from MPI functions
