+++ src/include/mpir_misc.h
@@ -87,6 +87,11 @@ int MPIR_Close_port_impl(const char *port_name);
 int MPIR_Open_port_impl(MPIR_Info * info_ptr, char *port_name);
 int MPIR_Cancel(MPIR_Request * request_ptr);
 
+int MPIR_Comm_join(int fd, MPI_Comm * intercomm);
+int MPIR_Comm_disconnect(MPIR_Comm * comm_ptr);
+int MPIR_Publish_name(const char *service_name, MPIR_Info * info_ptr, const char *port_name);
+int MPIR_Unpublish_name(const char *service_name, MPIR_Info * info_ptr, const char *port_name);
+
 /* Default routines for asynchronous progress thread */
 int MPIR_Init_async_thread(void);
 int MPIR_Finalize_async_thread(void);
