+++ src/mpi/spawn/comm_accept.c
@@ -23,14 +23,6 @@ int MPI_Comm_accept(const char *port_name, MPI_Info info, int root, MPI_Comm com
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Comm_accept
 #define MPI_Comm_accept PMPI_Comm_accept
-
-int MPIR_Comm_accept_impl(const char *port_name, MPIR_Info * info_ptr, int root,
-                          MPIR_Comm * comm_ptr, MPIR_Comm ** newcomm_ptr)
-{
-    return MPID_Comm_accept(port_name, info_ptr, root, comm_ptr, newcomm_ptr);
-}
-
-
 #endif
 
 /*@
